Time: 4 ms (52.60%), Space: 5.9 MB (26.75%) - LeetHub

+++ 69-sqrtx/69-sqrtx.cpp
@@ -14,6 +14,6 @@ public:
             else hi = mid - 1;
             
         }
-        return (int)ans;
+        return ans;
     }
 };
